# docs/openapi/commerce-api-v1.yaml
openapi: 3.1.0
info:
  title: AWJ Commerce Mobile API
  version: v1
  summary: Mobile/native-client commerce surface for AWJ tenants (`/commerce/v1`).
  description: >
    `/commerce/v1` is an additive, versioned surface for native mobile
    clients — a separate trust boundary from `/store/v1` (the web storefront)
    and `/api/v1` (the tenant-facing machine-to-machine Public API). It shares
    services with `/store/v1` (cart, checkout, catalog, orders) but never its
    controllers, and shares nothing with `/api/v1` beyond the error-code enum
    and general envelope conventions. See
    `docs/plans/commerce/PUBLIC_MOBILE_COMMERCE_API_V1_ARCHITECTURE.md`.


    ## Authentication tiers

    Every route requires a **store bearer token** (`Authorization: Bearer
    <key>`) — the same `ApiClient`/Sanctum credential `/api/v1` already uses,
    reused unmodified. It resolves the tenant and the tenant's single active
    `mobile` `SalesChannel`; it is never an end-customer's own session. Four
    route groups layer on top of that shared requirement:


    - **Read (optional customer context)** — store bearer only. An optional
      `X-Customer-Token` may additionally be presented (see below) to
      personalize cart/checkout resolution, but its absence never blocks a
      read.

    - **Write (optional customer context)** — same as Read, plus a stricter
      rate-limit class. Cart/checkout mutations live here.

...
      every other tenant's shared quota.

...


...
    of tier.


...

...


...

...


...

...
servers:
...
security:
  - storeBearer: []
paths:
  /storefront:
    parameters:
      - $ref: '#/components/parameters/RequestId'
    get:
      operationId: getStorefront
      tags: [Storefront]
      summary: Resolved store identity for the mobile channel
      x-auth-tier: read
      responses:
        '200':
          description: The tenant's store identity.
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StorefrontResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
  /checkout:
    parameters:
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/CartToken'
      - $ref: '#/components/parameters/OptionalCustomerToken'
    get:
      operationId: getCheckout
      tags: [Checkout]
      summary: Get the current checkout (shares Cart's X-Cart-Token identity)
      description: >
        Before `POST checkout` is ever called, this returns the checkout
        fields as `null`/zeroed alongside the real cart — never a 404 —
        exactly the same key set as a populated checkout.
      x-auth-tier: read
      responses:
        '200':
          description: The checkout (empty shape if none exists yet).
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
            X-Cart-Token: { $ref: '#/components/headers/XCartToken' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CheckoutResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
    post:
      operationId: startCheckout
      tags: [Checkout]
      summary: Start (or resume) checkout for the current cart
      x-auth-tier: write
      responses:
        '200':
          description: An already-open checkout resumed.
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
            X-Cart-Token: { $ref: '#/components/headers/XCartToken' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CheckoutResponse' }
        '201':
          description: A new checkout created.
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
            X-Cart-Token: { $ref: '#/components/headers/XCartToken' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CheckoutResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
  /addresses/{id}:
    parameters:
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/CustomerToken'
      - $ref: '#/components/parameters/IdPath'
    patch:
      operationId: updateAddress
      tags: [Addresses]
      summary: Update a saved address
      description: >
        Validates the **merged** (existing + patched) state, not the raw
        partial payload — a patch that flips `country` to `SA` or clears a
        required field cannot slip through.
      x-auth-tier: customer
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddressUpdate' }
      responses:
        '200':
          description: The updated address.
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AddressResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
    delete:
      operationId: deleteAddress
      tags: [Addresses]
      summary: Delete a saved address
      x-auth-tier: customer
      responses:
        '200':
          description: Deleted.
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeletedResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
components:
  securitySchemes:
    storeBearer:
      type: http
      scheme: bearer
      description: >
        The tenant-owned ApiClient/Sanctum key for this store's mobile
        client, identical to `/api/v1`'s own bearer mechanism.
  parameters:
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      description: Optional client-supplied correlation id, echoed back if valid; otherwise server-generated.
      schema: { type: string }
    CartToken:
      name: X-Cart-Token
      in: header
      required: false
      description: Opaque guest cart/checkout identity, issued in the response of the first cart mutation.
      schema: { type: string }
    CustomerToken:
      name: X-Customer-Token
      in: header
      required: true
      description: The customer's own `customer:access` Sanctum token (never the `Authorization` header).
      schema: { type: string }
    OptionalCustomerToken:
      name: X-Customer-Token
      in: header
      required: false
      description: >
        Optional on cart/checkout operations only (`EstablishCommerceCustomerContextIfPresent`):
        when presented alongside `X-Cart-Token`, the guest cart is claimed or
        merged into this customer's own identity (ADR-07), and this checkout
        completes as this customer's order rather than a guest one. Its
        absence never blocks the request — the same operation remains fully
        guest-usable.
      schema: { type: string }
...
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  headers:
    XRequestId:
      description: Correlation id for this response.
      schema: { type: string }
    XCartToken:
      description: Present when a new or rebound cart/checkout token must be persisted by the client.
      schema: { type: string }
  responses:
    Unauthenticated:
      description: Missing or invalid store bearer token (or, on customer-tier routes, X-Customer-Token).
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: >
        Resource not found, or a non-revealing 404 standing in for an
        ownership/publication failure (never distinguished from a genuinely
        missing id).
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Request failed validation.
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    TooManyRequests:
      description: Rate limit exceeded.
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Money:
      type: object
      additionalProperties: false
      required: [amount_minor, currency]
      properties:
        amount_minor: { type: integer, description: Minor units (e.g. halalas). Never a float. }
        currency: { type: string, description: 'ISO-4217 currency code.' }
    Error:
      type: object
      additionalProperties: false
      required: [error, meta]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
              enum: [internal_error, bad_request, not_found, method_not_allowed, validation_failed, unauthenticated, forbidden, tenant_context_required, client_inactive, insufficient_scope, rate_limited, idempotency_key_required, invalid_idempotency_key, idempotency_conflict, idempotency_in_progress, review_required, cart_merged]
            message: { type: string }
            details: {}
        meta:
          type: object
          required: [request_id]
          properties:
            request_id: { type: string }
    StorefrontResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [name]
          properties:
            name: { type: [string, 'null'] }
        meta: { $ref: '#/components/schemas/SuccessMeta' }
...
    SuccessMeta:
      type: object
      required: [request_id]
      properties:
        request_id: { type: string }
...
    CartItem:
      type: object
      additionalProperties: false
      required: [id, product_id, product_variant_id, variant_descriptor, product_name, unit_key, unit_name, quantity, unit_price, line_total, available]
      properties:
        id: { type: string, format: uuid }
        product_id: { type: [string, 'null'], format: uuid }
        product_variant_id: { type: [string, 'null'], format: uuid }
        variant_descriptor: { type: [string, 'null'] }
        product_name: { type: string }
        unit_key: { type: string }
        unit_name: { type: [string, 'null'] }
        quantity: { type: integer }
        unit_price: { $ref: '#/components/schemas/Money' }
        line_total: { $ref: '#/components/schemas/Money' }
        available: { type: boolean }
        personalization: { $ref: '#/components/schemas/LinePersonalization' }
    LinePersonalization:
      description: The customer's personalization values for a line (ADR-16). Present only on a personalized line.
      type: array
      items:
        type: object
        additionalProperties: false
        required: [key, label, label_en, value, value_label]
        properties:
          key: { type: string }
          label: { type: string }
          label_en: { type: [string, 'null'] }
          value: { type: string }
          value_label: { type: [string, 'null'] }
    Cart:
      type: object
      additionalProperties: false
      required: [status, items, subtotal, currency, has_unavailable_items]
      properties:
        status: { type: [string, 'null'] }
        items: { type: array, items: { $ref: '#/components/schemas/CartItem' } }
        subtotal: { $ref: '#/components/schemas/Money' }
        currency: { type: string }
        has_unavailable_items: { type: boolean }
...
    CheckoutAddress:
      type: object
      additionalProperties: false
      required: [country, region, city, district, street, building_no, additional_number, postal_code, notes]
      properties:
        country: { type: [string, 'null'] }
        region: { type: [string, 'null'] }
        city: { type: [string, 'null'] }
        district: { type: [string, 'null'] }
        street: { type: [string, 'null'] }
        building_no: { type: [string, 'null'] }
        additional_number: { type: [string, 'null'] }
        postal_code: { type: [string, 'null'] }
        notes: { type: [string, 'null'] }
    CheckoutPayment:
      type: object
      additionalProperties: false
      required: [payment_method_id, payment_method_name, method]
      properties:
        payment_method_id: { type: [string, 'null'], format: uuid }
        payment_method_name: { type: [string, 'null'] }
        method: { type: [string, 'null'], enum: [cod, pay_on_pickup, null] }
...
    Checkout:
      type: object
      additionalProperties: false
      required: [status, contact, delivery, payment, cart]
      properties:
        status: { type: [string, 'null'], enum: [active, ready, completed, expired, null] }
        contact:
          type: object
          additionalProperties: false
          required: [name, phone, email]
          properties:
            name: { type: [string, 'null'] }
            phone: { type: [string, 'null'] }
            email: { type: [string, 'null'] }
        delivery:
          type: object
          additionalProperties: false
          required: [method, amount, address]
          properties:
            method: { type: [string, 'null'], enum: [pickup, standard, null] }
            amount: { $ref: '#/components/schemas/Money' }
            address: { $ref: '#/components/schemas/CheckoutAddress' }
        payment: { $ref: '#/components/schemas/CheckoutPayment' }
        cart: { $ref: '#/components/schemas/Cart' }
    CheckoutResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/Checkout' }
        meta: { $ref: '#/components/schemas/SuccessMeta' }
...
    Address:
      type: object
      additionalProperties: false
      required: [id, label, recipient_name, phone, country, region, city, district, street, building_no, additional_number, postal_code, short_address, latitude, longitude, delivery_notes, is_default_shipping, is_default_billing, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        label: { type: [string, 'null'] }
        recipient_name: { type: string }
        phone: { type: string }
        country: { type: string }
        region: { type: [string, 'null'] }
        city: { type: string }
        district: { type: [string, 'null'] }
        street: { type: string }
        building_no: { type: [string, 'null'] }
        additional_number: { type: [string, 'null'] }
        postal_code: { type: [string, 'null'] }
        short_address: { type: [string, 'null'] }
        latitude: { type: [number, 'null'] }
        longitude: { type: [number, 'null'] }
        delivery_notes: { type: [string, 'null'] }
        is_default_shipping: { type: boolean }
        is_default_billing: { type: boolean }
        created_at: { type: [string, 'null'], format: date-time }
        updated_at: { type: [string, 'null'], format: date-time }
    AddressResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/Address' }
        meta: { $ref: '#/components/schemas/SuccessMeta' }
...
    AddressUpdate:
      type: object
      additionalProperties: false
      description: All fields optional — a partial patch, validated against the merged (existing + patched) state.
      properties:
        label: { type: [string, 'null'], maxLength: 255 }
        recipient_name: { type: string, maxLength: 255 }
        phone: { type: string, maxLength: 32 }
        country: { type: string, minLength: 2, maxLength: 2 }
        region: { type: [string, 'null'], maxLength: 255 }
        city: { type: string, maxLength: 255 }
        district: { type: [string, 'null'], maxLength: 255 }
        street: { type: string, maxLength: 255 }
        building_no: { type: [string, 'null'], maxLength: 32 }
        additional_number: { type: [string, 'null'], maxLength: 32 }
        postal_code: { type: [string, 'null'], maxLength: 32 }
        short_address: { type: [string, 'null'], maxLength: 8 }
        latitude: { type: [number, 'null'], minimum: -90, maximum: 90 }
        longitude: { type: [number, 'null'], minimum: -180, maximum: 180 }
        delivery_notes: { type: [string, 'null'], maxLength: 1000 }
        is_default_shipping: { type: boolean }
        is_default_billing: { type: boolean }
    DeletedResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [deleted]
          properties:
            deleted: { type: boolean }
        meta: { $ref: '#/components/schemas/SuccessMeta' }